requirements.txt
alembic==1.4.3
amqp==2.6.1
Flask==1.1.2
Flask-Admin==1.5.6
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
psycopg2-binary==2.8.6
pydantic==1.6.1
SQLAlchemy-i18n==1.0.3
SQLAlchemy-Utils==0.36.8
Werkzeug==1.0.1
